ai_service/requirements.txt
fastapi
uvicorn
vanna
psycopg2
python-dotenv
